# services/tracker/openapi.json
{
  "components": {
    "schemas": {
...
      "AgentDownloadURLResponse": {
        "properties": {
          "download_url": {
            "title": "Download Url",
            "type": "string"
          },
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "download_url",
          "expires_in"
        ],
        "title": "AgentDownloadURLResponse",
        "type": "object"
      },
      "AgentEntry": {
        "properties": {
          "last_modified": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Modified"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "AgentEntry",
        "type": "object"
      },
      "AgentsResponse": {
        "properties": {
          "agents": {
            "items": {
              "$ref": "#/components/schemas/AgentEntry"
            },
            "title": "Agents",
            "type": "array"
          }
        },
        "required": [
          "agents"
        ],
        "title": "AgentsResponse",
        "type": "object"
      },
...
      "BenchmarkServiceCatalogResponse": {
        "properties": {
          "services": {
            "items": {
              "$ref": "#/components/schemas/BenchmarkServiceEntry"
            },
            "title": "Services",
            "type": "array"
          }
        },
        "required": [
          "services"
        ],
        "title": "BenchmarkServiceCatalogResponse",
        "type": "object"
      },
      "BenchmarkServiceEntry": {
        "properties": {
          "auth_header_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auth Header Name"
          },
          "auth_secret_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auth Secret Name"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "name",
          "url"
        ],
        "title": "BenchmarkServiceEntry",
        "type": "object"
      },
      "BenchmarkServiceHealth": {
        "properties": {
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "healthy": {
            "title": "Healthy",
            "type": "boolean"
          },
          "latency_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Ms"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "name",
          "url",
          "healthy",
          "latency_ms"
        ],
        "title": "BenchmarkServiceHealth",
        "type": "object"
      },
      "BenchmarkServicesRequest": {
        "properties": {
          "services": {
            "items": {
              "$ref": "#/components/schemas/BenchmarkServiceEntry"
            },
            "title": "Services",
            "type": "array"
          }
        },
        "title": "BenchmarkServicesRequest",
        "type": "object"
      },
      "BenchmarkServicesResponse": {
        "properties": {
          "services": {
            "items": {
              "$ref": "#/components/schemas/BenchmarkServiceHealth"
            },
            "title": "Services",
            "type": "array"
          }
        },
        "required": [
          "services"
        ],
        "title": "BenchmarkServicesResponse",
        "type": "object"
      },
      "BenchmarkStatus": {
        "enum": [
          "IN_PROGRESS",
          "STOPPING",
          "STOPPED",
          "FINISHED",
          "ERROR"
        ],
        "title": "BenchmarkStatus",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RunAttemptsResponse": {
        "properties": {
          "attempts": {
            "items": {
              "discriminator": {
                "mapping": {
                  "error": "#/components/schemas/RunErrorTaskAttempt",
                  "evaluation": "#/components/schemas/RunEvaluationTaskAttempt",
                  "execution": "#/components/schemas/RunExecutionTaskAttempt"
                },
                "propertyName": "kind"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/RunErrorTaskAttempt"
                },
                {
                  "$ref": "#/components/schemas/RunEvaluationTaskAttempt"
                },
                {
                  "$ref": "#/components/schemas/RunExecutionTaskAttempt"
                }
              ]
            },
            "title": "Attempts",
            "type": "array"
          },
          "total_count": {
            "title": "Total Count",
            "type": "integer"
          }
        },
        "required": [
          "attempts",
          "total_count"
        ],
        "title": "RunAttemptsResponse",
        "type": "object"
      },
      "RunErrorTaskAttempt": {
        "properties": {
          "attempt_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attempt Id"
          },
          "attempt_number": {
            "title": "Attempt Number",
            "type": "integer"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "error_fingerprint": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Error Fingerprint",
            "type": "string"
          },
          "error_message": {
            "maxLength": 4000,
            "title": "Error Message",
            "type": "string"
          },
          "error_message_truncated": {
            "title": "Error Message Truncated",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "const": "error",
            "default": "error",
            "title": "Kind",
            "type": "string"
          },
          "status": {
            "const": "ERROR",
            "default": "ERROR",
            "title": "Status",
            "type": "string"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "attempt_id",
          "created_at",
          "error_message",
          "error_message_truncated",
          "error_fingerprint",
          "task_id",
          "attempt_number"
        ],
        "title": "RunErrorTaskAttempt",
        "type": "object"
      },
...
      "SingleBenchmarkResponse": {
        "description": "Single-run view: BenchmarkTableRow fields plus optional final_score.",
        "properties": {
          "agent_name": {
            "title": "Agent Name",
            "type": "string"
          },
          "cloudwatch_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cloudwatch Url"
          },
          "error_message": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "final_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Score"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "finished_tasks": {
            "title": "Finished Tasks",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "runtime": {
            "enum": [
              "managed",
              "legacy"
            ],
            "title": "Runtime",
            "type": "string"
          },
          "s3_bucket_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "S3 Bucket Url"
          },
          "started_at": {
            "title": "Started At",
            "type": "string"
          },
          "started_by_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started By Email"
          },
          "status": {
            "$ref": "#/components/schemas/BenchmarkStatus"
          },
          "task_state_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "default": {},
            "title": "Task State Counts",
            "type": "object"
          },
          "total_tasks": {
            "title": "Total Tasks",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "agent_name",
          "model",
          "started_at",
          "finished_at",
          "status",
          "total_tasks",
          "finished_tasks",
          "runtime"
        ],
        "title": "SingleBenchmarkResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "DescopeAccessKey": {
        "in": "header",
        "name": "x-api-key",
        "type": "apiKey"
      },
      "DescopeBearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/agents": {
      "get": {
        "description": "List agent zips under the org's S3 bucket.",
        "operationId": "list_agents_endpoint",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentsResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "DescopeBearer": []
          },
          {
            "DescopeAccessKey": []
          }
        ],
        "summary": "List Agents Endpoint"
      }
    },
    "/agents/{name}/download-url": {
      "get": {
        "description": "Return a 5-minute presigned URL to download agents/<name>.zip.",
        "operationId": "get_agent_download_url",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentDownloadURLResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "DescopeBearer": []
          },
          {
            "DescopeAccessKey": []
          }
        ],
        "summary": "Get Agent Download Url"
      }
    },
...
    "/benchmark-services": {
      "get": {
        "description": "Fetch catalog benchmark services visible to the caller.",
        "operationId": "catalog_benchmark_services",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BenchmarkServiceCatalogResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "DescopeBearer": []
          },
          {
            "DescopeAccessKey": []
          }
        ],
        "summary": "Catalog Benchmark Services"
      },
      "post": {
        "description": "Health-check the caller-provided benchmark services with concurrent pings.",
        "operationId": "list_benchmark_services",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BenchmarkServicesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BenchmarkServicesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "DescopeBearer": []
          },
          {
            "DescopeAccessKey": []
          }
        ],
        "summary": "List Benchmark Services"
      }
    },
...
    "/benchmarks/{benchmark_id}": {
      "get": {
        "description": "Fetch a single benchmark with task counts + final score for the SingleRun page.",
        "operationId": "get_single_benchmark",
        "parameters": [
          {
            "in": "path",
            "name": "benchmark_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Benchmark Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SingleBenchmarkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "DescopeBearer": []
          },
          {
            "DescopeAccessKey": []
          }
        ],
        "summary": "Get Single Benchmark"
      }
    },
    "/benchmarks/{benchmark_id}/attempts": {
      "get": {
        "description": "Return every persisted task outcome in a run, newest first.",
        "operationId": "get_run_attempts",
        "parameters": [
          {
            "in": "path",
            "name": "benchmark_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Benchmark Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunAttemptsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "DescopeBearer": []
          },
          {
            "DescopeAccessKey": []
          }
        ],
        "summary": "Get Run Attempts"
      }
    },
...
  }
}